Fix Article Paths to reflect Shopify
